Optionally send slack notifications when the build is fixed.
Adds a new configuration option, :first-success, that, when set to true, will
determine whether the current successful build is the first success after a
failure. If so, it will send a notification to the configured slack channel.

Adds tests for slack notifications to the existing tests of the main function.

Fixes #44.