Handle go-junit-report XML.
This tool produces a file with multiple test suite reports in it, whereas it's
usually the case that we encounter a single testsuite per XML file. We now
handle this scenario, as well as cases where the test report may not contain
information about how many tests were skipped, or how many errors there were.