Split wrap-build-meta and reorder some middleware
I believe this was the cause for the errors we saw after 1.3.6 was
deployed.  Essentially, there are steps that presume that there is a
checked-out copy of the code and the change to infer the branch from
the job name introduced a chicken and egg problem b/c it needed
wrap-build-meta to run in order to get the branch name in order to
clone but the second half of wrap-build-meta (now called
wrap-last-success) expected the repo to have already been cloned.