Add a little debugging output
I don't want to add much b/c, in theory, I can search all of this in
ES but in the currently failing case it appears that something odd is
being passed in for "local" (a Jenkins URL) and I wanted to log a bit
more to be sure.  Also I cannot find this test output on build-stats.