Fetch extra commits when the clone was shallow
Shallow clones make it tricky to check out a particular commit unless
that commit happens to be HEAD.  So, in the case of a shallow clone,
fetch 6 months worth of commits with the hopes that will cover the
commit to be tested.

This is a bit of a tradeoff between using '--unshallow' and paying the
price for a full clone and limiting how far back in time we can go for
fetching.