use multipart/mixed along with a multipart/alternative element
Previously there were often 3 parts in the multipart/alternative
section (attachment, plain text, html).  This displayed correctly in
several email clients but failed to render attachments in Thunderbird.
It appears that Thunderbird is correct, according to the spec.
Instead, the proper solution is to have a multipart/mixed email with
attachments and a multipart/alternative element with the txt/html
parts.