Quiet down some test output
runbld captures the stdout and stderr of the wrapped process and later
forwards that text to its own out and err streams.  This is essential
when running but produces spurious output when testing.  This change
allows the tests to rebind the out and err streams to a writer that
writes those to the test.log file.